cleaned up the backend

+++ server/init/initializeDatabase.js
@@ -167,6 +167,9 @@ async function clearSchedule(teacher) {
     },
   });
   // teacherSchedule.id may be null here
+  if(!teacherSchedule){
+    return;
+  }
   await prisma.scheduledClass.deleteMany({
     where: {
       scheduleId: teacherSchedule.id,
